refactor(details): extract and inject CountryDetailsViewModel; remove UIKit integration from SwiftUI
- Introduced CountryDetailsViewModel with flagLoadingState, and shareText
- Updated CountryDetailsView to accept and bind injected viewModel
- Replaced UIKit-based share logic with SwiftUI ShareLink
